update api san pham, khuyen mai & danh muc
