+++ src/main.tsx
@@ -1,15 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.tsx'
+import Providers from './pages/providers/providerRoute.tsx'
 import 'antd/dist/reset.css'
 import { ToastContainer } from 'react-toastify'
 import { BrowserRouter } from 'react-router'
-import { persistor, store } from './redux/store.ts'
+import { persistor, store } from './pages/store/redux/store.ts'
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-
 
 const queryClient = new QueryClient()
 
@@ -19,7 +18,7 @@ createRoot(document.getElementById('root')!).render(
       <Provider store={store}>
         <PersistGate loading={null} persistor={persistor}>
           <QueryClientProvider client={queryClient}>
-            <App />
+            <Providers />
             <ToastContainer
               position="top-center"
               autoClose={5000}
